+++ package.json
@@ -18,6 +18,7 @@
     "expo": "~50.0.6",
     "expo-constants": "~15.4.5",
     "expo-dev-client": "~3.3.8",
+    "expo-image": "~1.10.6",
     "expo-image-picker": "~14.7.1",
     "expo-location": "~16.5.3",
     "expo-status-bar": "~1.11.1",
@@ -27,6 +28,7 @@
     "react": "18.2.0",
     "react-native": "0.73.4",
     "react-native-bouncy-checkbox": "^3.0.7",
+    "react-native-date-picker": "^4.4.0",
     "react-native-dropdown-select-list": "^2.0.5",
     "react-native-elements": "^3.4.3",
     "react-native-gesture-handler": "~2.14.0",
@@ -34,7 +36,7 @@
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
     "react-native-url-polyfill": "^2.0.0",
-    "expo-image": "~1.10.6"
+    "@react-native-community/datetimepicker": "7.6.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
